Refactor Metadata API a little and reformat code
All metadata processing is now done in the metadata package. All code
and tests were updated to reflect this. I'm still
not super happy with the API, but I will wait until the Free refactor to
clean it further.

Fixed a bug where Stage interval metadata was parsed as an Int rather
than a Double.

Updated dependencies and plugins to the latest version.

+++ io/src/main/scala/llsm/io/metadata/WaveformMetadata.scala
@@ -27,7 +27,7 @@ object WaveformMetadata {
   case class Excitation(channel: Int,
                         filter: String,
                         laser: Int,
-                        power: Int,
+                        power: Double,
                         exposure: Double)
   case class Cycle(mode: String)
   case class ZMotion(mode: String)
